refactor(rwa): move country module tests into sibling files
Align the country compliance modules with the repository's preferred test
layout by extracting inline test modules into dedicated test.rs files.
